+++ docs/CHANGES.md
@@ -1,6 +1,7 @@
 2017-06-06  Release 1.81.17
 
   * Bug fix: Don't crash when the OpenVAS scan report is invalid or empty.
+  * Bug fix: Don't push a canvas in empty dashboard table cells.
   
 
 2017-06-06  Release 1.81.16